# BackendServices/matrimonialwebapi/src/MatrimonialApi/Controllers/FavouriteApi.cs
/*
 * Matrimonial API - OpenAPI 3.0
 *
 * Design and definition of Matrimonial APIs created for practice and teaching
 *
 * OpenAPI spec version: 1.0.11
 * Contact: floatingrays@gmail.com
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */
using System;
using System.Collections.Generic;
using Microsoft.AspNetCore.Mvc;
using Swashbuckle.AspNetCore.Annotations;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using IO.Swagger.Attributes;
using IO.Swagger.Security;
using Microsoft.AspNetCore.Authorization;
using IO.Swagger.Models;

namespace IO.Swagger.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    [ApiController]
    public class FavouriteApiController : ControllerBase
    { 
        /// <summary>
        /// Add a favourite profile
        /// </summary>
        /// <remarks>Add a favourite profile</remarks>
        /// <param name="body">Update a Profile</param>
        /// <param name="xRequestAuth"></param>
        /// <response code="200">Successful operation</response>
        /// <response code="400">Invalid input</response>
        /// <response code="422">Validation exception</response>
        /// <response code="500">An error occured while processing the request.</response>
        /// <response code="0">Default error</response>
        [HttpPost]
        [Route("/api/favourite")]
        [Authorize(AuthenticationSchemes = BearerAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("AddFavourite")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<Favourite>), description: "Successful operation")]
        public virtual IActionResult AddFavourite([FromBody]Favourite body, [FromHeader][Required()]string xRequestAuth)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<Favourite>));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 422 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(422);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "[ {\n  \"personId\" : 0,\n  \"matchid\" : 6\n}, {\n  \"personId\" : 0,\n  \"matchid\" : 6\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<Favourite>>(exampleJson)
                        : default(List<Favourite>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Delete a favourite profile
        /// </summary>
        /// <remarks>Delete a favourite profile</remarks>
        /// <param name="xRequestAuth"></param>
        /// <param name="profileId"></param>
        /// <param name="favouriteIdToBeDelete"></param>
        /// <response code="200">Successful operation</response>
        /// <response code="400">Invalid input</response>
        /// <response code="422">Validation exception</response>
        /// <response code="500">An error occured while processing the request.</response>
        /// <response code="0">Default error</response>
        [HttpDelete]
        [Route("/api/favourite/{profileId}")]
        [Authorize(AuthenticationSchemes = BearerAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("DeleteFavourite")]
        public virtual IActionResult DeleteFavourite([FromHeader][Required()]string xRequestAuth, [FromRoute][Required]int? profileId, [FromQuery][Required()]int? favouriteIdToBeDelete)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200);

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 422 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(422);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);

            throw new NotImplementedException();
        }

        /// <summary>
        /// Get favourites for a profile
        /// </summary>
        /// <remarks>Get favourites for a profile</remarks>
        /// <param name="xRequestAuth"></param>
        /// <param name="profileId"></param>
        /// <response code="200">Successful operation</response>
        /// <response code="400">Invalid input</response>
        /// <response code="422">Validation exception</response>
        /// <response code="500">An error occured while processing the request.</response>
        /// <response code="0">Default error</response>
        [HttpGet]
        [Route("/api/favourite/{profileId}")]
        [Authorize(AuthenticationSchemes = BearerAuthenticationHandler.SchemeName)]
        [ValidateModelState]
        [SwaggerOperation("GetFavourites")]
        [SwaggerResponse(statusCode: 200, type: typeof(List<ShortProfile>), description: "Successful operation")]
        public virtual IActionResult GetFavourites([FromHeader][Required()]string xRequestAuth, [FromRoute][Required]int? profileId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(List<ShortProfile>));

            //TODO: Uncomment the next line to return response 400 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(400);

            //TODO: Uncomment the next line to return response 422 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(422);

            //TODO: Uncomment the next line to return response 500 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(500);

            //TODO: Uncomment the next line to return response 0 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(0);
            string exampleJson = null;
            exampleJson = "[ {\n  \"firstName\" : \"firstName\",\n  \"lastName\" : \"lastName\",\n  \"cityName\" : \"cityName\",\n  \"stateName\" : \"stateName\",\n  \"imageURL\" : \"imageURL\",\n  \"occupationName\" : \"occupationName\",\n  \"middleName\" : \"middleName\",\n  \"id\" : 0\n}, {\n  \"firstName\" : \"firstName\",\n  \"lastName\" : \"lastName\",\n  \"cityName\" : \"cityName\",\n  \"stateName\" : \"stateName\",\n  \"imageURL\" : \"imageURL\",\n  \"occupationName\" : \"occupationName\",\n  \"middleName\" : \"middleName\",\n  \"id\" : 0\n} ]";
            
                        var example = exampleJson != null
                        ? JsonConvert.DeserializeObject<List<ShortProfile>>(exampleJson)
                        : default(List<ShortProfile>);            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}