# BackendServices/matrimonialwebapi/src/MatrimonialApi/wwwroot/swagger-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "title" : "Matrimonial API - OpenAPI 3.0",
    "description" : "Design and definition of Matrimonial APIs created for practice and teaching",
    "contact" : {
      "email" : "floatingrays@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:3002/api"
  } ],
...
  "paths" : {
...
    "/role" : {
      "summary" : "Manage roles",
      "description" : "Manage roles",
      "get" : {
        "tags" : [ "Role" ],
        "summary" : "Get all roles",
        "description" : "Get all the roles",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Role"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false,
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/role/{roleId}" : {
      "summary" : "Get a role",
      "description" : "Get a role",
      "get" : {
        "tags" : [ "Role" ],
        "summary" : "Get a existing role",
        "description" : "Get a existing role",
        "parameters" : [ {
          "name" : "roleId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Request-Auth",
          "in" : "header",
          "description" : "Authorization token value",
          "required" : false,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Role"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "deprecated" : false
      }
    },
...
    "/favourite/{profileId}" : {
      "summary" : "Get a favourite",
      "description" : "Get a favourite",
      "get" : {
        "tags" : [ "Favourite" ],
        "summary" : "Get favourites for a profile",
        "description" : "Get favourites for a profile",
        "operationId" : "GetFavourites",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "description" : "Autorization"
          }
        }, {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ShortProfile"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "Favourite" ],
        "summary" : "Delete a favourite profile",
        "description" : "Delete a favourite profile",
        "operationId" : "DeleteFavourite",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string",
            "description" : "Autorization"
          }
        }, {
          "name" : "profileId",
          "in" : "path",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "favouriteIdToBeDelete",
          "in" : "query",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation"
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
    "/favourite" : {
      "summary" : "Add or remove a favourite",
      "description" : "Add or remove a favourite",
      "post" : {
        "tags" : [ "Favourite" ],
        "summary" : "Add a favourite profile",
        "description" : "Add a favourite profile",
        "operationId" : "AddFavourite",
        "parameters" : [ {
          "name" : "X-Request-Auth",
          "in" : "header",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "Update a Profile",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Favourite"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Favourite"
                  },
                  "x-content-type" : "application/json"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input"
          },
          "422" : {
            "description" : "Validation exception"
          },
          "500" : {
            "description" : "An error occured while processing the request."
          },
          "default" : {
            "description" : "Default error"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Role" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "roleName" : {
            "type" : "string"
          },
          "permissionName" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "example" : {
          "roleName" : "roleName",
          "id" : 0,
          "permissionName" : [ "permissionName", "permissionName" ]
        }
      },
      "ShortProfile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "middleName" : {
            "type" : "string"
          },
          "imageURL" : {
            "type" : "string"
          },
          "cityName" : {
            "type" : "string"
          },
          "stateName" : {
            "type" : "string"
          },
          "occupationName" : {
            "type" : "string"
          }
        },
        "example" : {
          "firstName" : "firstName",
          "lastName" : "lastName",
          "cityName" : "cityName",
          "stateName" : "stateName",
          "imageURL" : "imageURL",
          "occupationName" : "occupationName",
          "middleName" : "middleName",
          "id" : 0
        }
      },
      "Favourite" : {
        "type" : "object",
        "properties" : {
          "personId" : {
            "type" : "integer"
          },
          "matchid" : {
            "type" : "integer"
          }
        },
        "example" : {
          "personId" : 0,
          "matchid" : 6
        }
      },
...
    },
    "securitySchemes" : {
      "BasicAuth" : {
        "type" : "http",
        "scheme" : "basic"
      },
      "BearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      },
      "SSO_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://localhost:3002/oauth/authorize",
            "scopes" : {
              "write:pets" : "modify data",
              "read:pets" : "read data"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}